fix service browsing
We were incrementing req->search_uuid every time we call back into
browse_cb() -- even when we'd already finished going through the
uuid_list[] and were supposed to be handling req->uuids.

So for each entry in req->uuids, we'd skip a zero after uuid_list[] and
go trawling through random memory, treating it as more UUIDs to search
for.