Check for NULL return of adapter_find_device
In theory the device might have been removed if the D-Bus client exited before
the remote side accepted the pairing. This is just a short term fix since
security.c is still storing the link key. In the long run we should move the
link key storing to dbus-hci.c and only do it when we get auth complete or
connect complete.