Move dev->sink initialization as acceptor to correct place
We can't initialize dev->sink in the AVDTP server callback since at this point
we don't know if the remote device will be acting in audio source or audio sink
role. So, do the initialization when we get the SetConfiguration command since
then we can check the type of the SEP that was selected.