Check for headset connected state when unregistering
In theory a headset could be in a connected state when the driver for it gets
unregistered. Do proper cleanup in this case (call headset_set_state(hs,
DISCONNECTED)). Currently this happens even in practice due to a bug where we
incorrectly assume that a device has removed a profile when in fact it's just
hiding it while it's connected.