Fix the service discovery-upon-pairing logic
The original code seems to try to handle the situation of two opposite directed
pairing attempts (remote side starts dedicated bonding with us but someone
calls CreatePairedDevice on our side at the same time). I'm not sure how likely
this actually is or if it can even succeed, but the existing logic in the code
was nevertheless wrong. After this patch is at least in theory makes sense to
me (and is better commented if I forget what I was thinking when I wrote it).