Don't remove service discovery timer even if device disconnects
Since it's important for us to find out what profiles a device has in order to
populate the device object with the correct D-Bus interfaces insist on doing
SDP even if got disconnected from the device. Since the timer is only 2 seconds
the chances are that device will still be around and the connection will be
successfull. This situation can happen if the remote device acts as an
initiator for dedicated bonding and brings down the ACL before our timer goes
off.