Clean up properly if stream starting fails
Remove the stream state callback if stream start fails. Without this patch the
callback could still be called after a unix client has disconnected which in
turn would cause a double-free to occur. There is code to remove the callback
also in client_free() but it won't work since we remove any references to the
session and stream in the resume callback.