Limit the amount of connected HSP/HFP devices and make it configurable
It's very tricky to get the multiple connected HSP/HFP devices use case right.
E.g. in the case of in-band ringtone the specification mandates us to open SCO
to the HFP before sending RING indications. So, in the case of multiple
connected devices we would have to open multimple simultaneous SCO connections
which simply isn't a wise thing to do.