obexd: Change xfer callback and dc_cb order in the event of a disconnection
This is mainly working around a bug in the gwobex API. Since it doesn't have
proper refcounting the xfer callback can free the gwobex ctx and so the second
callback fails. It could of course happen in the other direction too (dc_cb
calls gw_obex_close) but since no known code is using dc_cb it's at least a
little safer this way.