obexd: Try to reopen the tty once if the connection gets dropped
This ensures that we're back monitoring the tty in case the client sends
a OBEX Disconnect which in the current code always does a
OBEX_TransportDisconnect. This patch is also preparation for the next
step where we should be able to re-initialize the tty through some
external trigger (e.g. a SIGUSR1 signal).