ix adapter initialization logic
The original variable that was used for the initial DEVDOWN ioctl when
OffMode=DevDown and the stored mode is "off" can't really be used elsewhere
since it gets reset after calling the ioctl even though e.g. drivers haven't
been probed yet. This patch renames the old variable (to first_up) and adds a
new one for tracking when the adapter initialization has fully been completed.