Get rid of the first_up variable and always probe drivers on startup
This patch reorganizes the adapter_up function logic a little bit so
that we probe all drivers even if we are going to do a DEVDOWN ioctl
afterwards due to a previously stored off mode (in the OffMode=DevDown
case). An extra benefit is that we don't need the second init tracking
variable in the adapter struct anymore.