Allow FindDevice to find temporary devices
E.g. in the case of an incoming object push request we may not be paired
with the remote device and therefore the temporary flag will be set.
However, if there's a "receive file" UI active it will most likely want
to show at least the remote device name and therefore it should be
possible to use FindDevice.