Don't auto-disconnect explicitly created A2DP connections
This patch makes sure that explicitly created connections (remotely
initiated or using AudioSink.Connect()) stay up until explictly
disconnected. The control for this needs to be in the stream timer and
not the disconnect timer since our AVDTP state machine always tries to
keep one stream in configured state (i.e. it should be ok to always
disconnect if we stay too long in IDLE state).