Set avdtp session watch to low priority.
The session watch should be low priority because on old kernels (prior
to DEFER_SETUP). This is needed for stream Open & Start because the
connection is automatically accepted and makes the remote device to
sending the Start command imediately. Since both the stream transport
connection and the Start command arrive so close to each other the
kernel might wake up bluetoothd at the same time for both.