Replace asound.conf with a more full-fledged bluetooth.conf
The new fragment will introduce device strings that can be given the BT
address of the headset as parameter. Also, they integrate 'plug' so that
they can be used with any application.

This replaces the old 'asound.conf' file which was much more limited.
Also since we now install 'bluetooth.conf' to /etc/alsa it is not a good
idea to simply call that file 'asound.conf'.

Please note that this will install the config fragment but not actually
enable it. For that some minor changes to the /etc/asoundrc as shipped
by the distro are necessary. It's up to the distributions to do this.
How that works in explained in the header of bluetooth.conf.