Handle case where inquiry is active when bluetoothd starts up
If inquiry is active when bluetoothd starts up, it will get an error
from the HCI layer when it tries to execute the StartDiscovery dbus
call.  This error in turn will prevent the list of devices which have
been discovered and reported over dbus from being cleared, which will
make many devices undiscoverable.

This patch handles the case where the adapter is in standard inquiry
mode.  It checks for the HCI_INQUIRY status bit at startup, and if set,
issues an inquiry_cancel call.