Cancel periodic inquiry at startup
If periodic inquiry is active when bluetoothd starts up, it will get an
error from the HCI layer when it tries to execute the StartDiscovery
dbus call.  This error in turn will prevent the list of devices which
have been discovered and reported over dbus from being cleared, which
will make many devices undiscoverable.

This patch handles the case where the adapter is in periodic inquiry
mode at startup.  It issues a command to cancel periodic inquiry at
startup if the adapter is not known to be in standard inquiry mode,
hasn't been initialized yet and not DOWN when bluetoothd started.