Fix local class of device race conditions upon bluetoothd startup
The bluetoothd startup sequence is quite fragile and easily susceptible to race
conditions. The problematic variable that this patch fixes is
adapter->dev.class which depends on at least one write_class_of_dev_complete
event. If e.g. a write_scan_complete comes before this the
set_limited_discoverable function would attempt to set the wrong class of
device (essentially leaving out the service class bits completely).

What this patch does is adds tracking to when bluetoothd is still starting up
and caches the changes to the service classes right until the very moment when
it enters the main loop.