Don't force SCO up in the case of an incoming call
In principle HFP 1.5 requires us to bring SCO up in the case of an
incoming call (before we start sending RING indications). However, it's
not safe to force it up ourselves in bluetoothd since there could be an
A2DP stream active which needs to be suspended first and the telephony
subsystem might want to do some extra initialization before enabling
SCO. This patch basicly leaves the responsibility of bringing up SCO to
an external entity but still leaves the hs->pending_ring flag on so that
we know to start sending RING indications when SCO finally goes up.