obexd: Bring back tty reopen attempt on POLLHUP
It turns out that to get PC Suite to work properly we need to keep the
POLLHUP notifications in the kernel and have userspace retry opening the
tty on such an event. This commit essentially reverts commit
50ca8715834cbcdbf15aac3abfd75d287ebacede and adapts it to the current
state of obex.c (several things have changed so a clean revert isn't
possible).