Fix service classes race condition upon starting bluetoothd
During initial startup if InitiallyPowered=false we might not get the
"write class of device complete" HCI event before bluetoothd puts the
adapter DOWN. In this case the correct class never gets written to storage
but stays in adapter->svc_cache instead. This patch makes sure that the
right class is always set in adapter_up().