Map AVRCP Play and Pause to separate dedicated key codes
This change is needed because of some more advanced devices that have
special assumptions about what happens when they send a AVRCP Play and
Pause.

As an example, one carkit was found that didn't have separate play and
pause buttons but it had a "mute" button. This mute button caused the
device to send a AVRCP Pause and mute its audio stream. In the case that
the device is out of sync with the player status on the BlueZ side this
Pause button press would previously have been mapped to KEY_PLAYPAUSE
which in turn would cause the player to resume playing if it was already
paused. After this, selecting "unmute" on the car kit would cause it to
send a AVRCP Play and unmute its stream. However, since the BlueZ side
player would already be playing the KEY_PLAYPAUSE would get interpreted as
a pause and so there would again be no sound heard in the car kit.

Having separate dedicated key codes for AVRCP Play and Pause fixes this
issue.