Fix race condition on device removal.
There is a race condition which can cause the device object to be recreated
after being removed. The race condition can be triggered by a response that
comes after the cancel but if the device object is no longer available it is
interpreted as a new device and as such a new object is created. To fix this
race we have to wait for the device to disconnect and only then return.