Fix pairing process even if the ACL gets dropped before successful SDP
CreatePairedDevice should succeed if the pairing succeeds, but return only
after the SDP discovery is complete (either a success or a failure). Right
now the logic of the temporary flag and CancelDeviceCreation was such that
the device would be removed in the case of a successful pairing if the SDP
part failed.

This patch fixes the issue by clearing the temporary flag imediately after
a successful pairing but still allows CancelDeviceCreation to be called
after the pairing is complete but SDP isn't (i.e. Create*Device hasn't
returned anything yet). A new helper function for CancelDeviceCreation
called device_is_creating was added to allow easy checking of an ongoing
device creation.