Fix multiparty bit value when doing a conference call
There's no signal from CSD to indicate that the multiparty call status has
changed. The only way to figure out the multiparty status of a call is by
parsing the reply to the GetCallInfoAll method. So, every time a call goes
active or to hold and there is more than one call ongoing use
GetCallInfoAll to update the multiparty bit (and other properties that
come with this method call).