Add support for Call.CreateRequested signal
The HFP specification requires the AG to enable the "callheld" indicator
with the value 1, i.e. "Call is placed on hold or active/held calls
swapped" when there is an active call and a second outgoing call is
attempted which causes the active call to go on hold. However, with the
current csd and its underlying modem the existing call is set to hold
before any new call creation attempt is indicated upwards in the stack.
This means that we don't know about the second call just with the help of
the CallStatus signal (which maps directly to the modem state) but a new
signal, CreateRequested, is needed.

Since the create request could fail before any call object leaves the IDLE
state or the modem could get reset (in which case a CreateFailed signal
doesn't help) a five second timer is used to clear the call creation
information if the active call doesn't transition to the HOLD state before
that.