Fix three-way calling indicator order
MSC 4.26 and the text below it in the HFP 1.5 specification conflict with
each other with respect to the order of the callheld and callsetup
indicators. Errata 2459 clarifies the situation by explaining that the
text is correct and the MSC incorrect. Additionally the errata says that
the callheld indicator should be set to 1 just prior to indicting that the
the new call is active and the call setup procedure complete. This patch
updates the Maemo telephony driver logic accordingly.