Fix maemo telephony initialization to wait for correct modem state
It seems the most reliable way to know that CSD methods will succeed is to
listen for SSC modem state signals and only try to initialize our call
states when the modem is ready. However, since we need to have basic HFP
for VoIP even when the modem is not working we call telephony_ready_ind
already in telephony_init.