Fix org.bluez.Audio state tracking
When either Headset or AudioSink goes to connected Audio should go to
connected as long as the other profile is already connected. This patch
fixes the "other profile state" check to connected or any higher state
(e.g. playing) instead of strictly only testing for connected.

This bug could be triggered when the remote device is acting as initiator
and is very fast in bringing one profile into some state higher than
connected (before the other profile manages to become connected).