Use an idle callback for trusted device authorization
The caller of btd_request_authorization could assume that it's
callback will be called in a separate mainloop iteration as opposed to
from within the actual btd_request_authorization function. It's therefore
better to use an idle callback to make the trusted device callback
behavior similar to that of untrusted devices.