Extend read_link_key to be used as a simple boolean test
In some cases we just want to know if there's a link key or not but not
actually do anything with the key. This patch makes it possible to use the
existing read_link_key storage function to be able to check for the
existence of the link key without actually returning the key to the
calling function.