Implement caching of remote host features
Devices with SSP but no EIR requires looking at the remote host features
to get a correct value for the LegacyPairing property. Requesting a
remote name gives these features for free but if the name is already
cached we don't get them. This patch implements the caching part for the
features and a subsequent patch will actually make use of the cached
value for the device discovery process.