Add 0xff (any object type) to sdptool OPUSH supported formats list.
This fixed an issue with sending an image from OSX. OSX would show a
warning that the bluez stack does not support this file type, but offer
to send it anyway (which works). This change to advertise support for
all object types will remove this warning.

The list of file types that can be advertised in the SDP record is very
limited anyway. It is best to list that we support all file types here,
and then filter at the OPP/OBEX connection instead.

For reference:
0x01 = vCard 2.1
0x02 = vCard 3.0
0x03 = vCal 1.0
0x04 = iCal 2.0
0x05 = vNote
0x06 = vMessage
0xFF = any type of object.