Send abort on while cancelling a2dp setup
This should guarantee that any outstanding command in the queue is
cancelled, accourding to the avdtp spec:

"6.16 Abort
Because AVDTP is typically used on unreliable channels, signaling messages
can be lost due to L2CAP Flush Timeouts. To react to inconsistencies
between the INT and ACP states, the Abort Command may be used."

It also says abort can be send regardless of the state even in case of
packet fragmantation:

"Note that it is permitted to breakdown a packet sequence (command or
response message) to send an Abort command message in the same direction.
The remaining packets consecutive to an aborted message shall be flushed by
the sender."