Handle connection ending events on higher layers first
When sockets are notified simultaneously of a disconnection or
connection failure, the sockets on the higher levels should be allowed
to handle the disconnect event before cleaning up the connection at
sockets on the lower level (HCI).

The issue behind creating the patch was that on a failed connection
attempt hcid_dbus_conn_complete could get called before the SDP
callbacks, which lead to the latter not being called.

Lowers the priority of the HCI socket to trigger it after higher level
sockets such as L2CAP and RFCOMM.