Return error properly on CreateDevice
Adds a check in function search_cb for whether the initiating DBus
call was CreateDevice or CreatePairedDevice. Previously both calls
triggered a positive reply regardless of the error code, but in case
of CreateDevice any error should be propagated in the DBus reply.

Modifies the store_profiles function to return immediately for a
temporary device.