Fix Device.SetProperty("Trusted",... always emit PropertyChanged
Currently any change done to Trusted property generates a PropertyChanged
since no verification is done prior to apply the new value which may
generate an unnecessary write to the storage and extra D-Bus traffic.

To fix this trusted is now loaded from storage and cached in memory when
the device is created, which latter is used to make sure the value has
really changed avoiding storage reads to check current value.