obexd: Fix dealing with large files
In C, the result of an operation has always the size of the largest
operand, in this case both operands are unsigned longs, which was
causing the operation to overflow on 32bit machines.

The solution is to force an operand to be of a size that would not
overflow and to store the result somewhere with a safe size.