Fix DiscoverServices not retrieving any records
In the case that device->tmp_records was already set, req->records was
being set to NULL and so causes the empty array to be returned.

To fix that tmp_records now always has the latest discovered records and
it is used when repling to DiscoverServices.

Thanks to Tim Renouf <tpr.vger@botech.co.uk> for figuring out where was
the problem.