Fix handling of incoming call indication when in WAITING state
Some headsets expect incoming call indication before they can send ATA
command. In current implementation of telephony we do not send outband
notyfication about incoming call when call previously was in WAITING
state. And without that kind of notyfication BH-108 and BackBeat 903
headset cannot send ATA command - user is unable to answer to that call
from headset.

Added handling for that case in func handle_call_status(). When phone now
is in CSD_CALL_STATUS_MT_ALERTING and previously was in
CSD_CALL_STATUS_WAITING that means we need to send incoming call
notyfication (it was not sent earlier because new call was waiting for end
of active call). This will send RING commands to headset if it is needed.