obexd: Fix issues caused by multiple dependencies on phonebook.c
Previously only PBAP depended on phonebook.c. Now IrMC Sync also depends
on it. That means that phonebook_init and phonebook_exit need to be able
to handle multiple calls or else memory leaks and double-free's are
likely. This patch adds safeguards for this to the _init and _exit
functions that need it.