Store new local name before writing it to adapter
Store new local name to the file system and emit property changed
before actually writing the name to the adapter.

The reason for the patch is the scenario were an application sets
the name and then immediately turns off the adapter. Previously the
requested name change might not have been properly applied in that
case. The new name was not written to storage until it had been
read from the adapter and the adapter could have been turned off
before the read command was completed.

The patch avoids writing the name twice by using a boolean variable in
the btd_adapter struct. The name may be changed by using for example
hciconfig and in that case the name needs to be updated to storage after
reading.