Fix ATT server read_by_group not respecting Secondary services
According to the definition of ending of a service[1], the service
must end at the next service declaration, that includes a Secondary
Service.

There's one issue that I did not find a definitive answer in the spec,
in the case that the end handle of the request is less than the End
Group handle of the service, what should be put in the response?

[1] "The service definition ends before the next service declaration
or after the maximum Attribute Handle is reached."