obexd: Change default type of call from call log
Previously after downloading call log all calls had HOME type as default.
Now this type is OTHER - it will be used when connection is made
with number which is out of phonebook. To solve this issue extending
number of columns and queries of database was needed especially
(one field in database query was added).