Fix problem with operator name length
To follow the HFP specification we must ensure that network operator name
will have at maximum 16 characters. In current implementation we do not
check this constraint. This patch fix this. Now operator name with more
than 16 characters will be truncated to 16 characters.

Note: the csd API in question guarantees ASCII strings so there's no
risk of having a multi-byte character getting split on both sides of the
16th byte.