Fix accessing freed memory
Crash is due to some junk characters in the device name.
The device name is copied from a string variable after freeing the memory
it points to, so there were some junk characters in it which was the reason
for the crash.

Fixing the issue by freeing the memory in string variable after
device name is copied.