audio/device.h: use lower case in profile UUIDs
BlueZ seems to use lower case when emitting signals containing
UUIDs retrived from a remote device.

The definitions affected by this patch look to be used for example
if the remote connects an audio profile which BlueZ has not
discovered (perhaps it had not yet browsed the service records of
the remote).

In that case the format of the stored and emitted UUID should be
consistent with the UUIDs received during service discovery.