Fix A2DP play when sample rate gets changed
Playing a 48KHz wav file immediately after playing a 44.1K file (or
vice-versa) with aplay will fail. After this failure no a2dp stream will
play. The problem is that after determining the current avdtp stream
doesn't match the new configuration the stream is closed and a
reconfiguration is attempted but fails because setup->rsep is NULL. This
patch below fixes the reconfiguration.