Simplify D-Bus connection handling and move it to main.c
The reconnection logic for the system bus doesn't make much sense. If
the system bus daemon dies so does most useful system functionality.
Since the code for handling D-Bus setup and cleanup is very simple after
this removal it makes sense to move it to main.c and make the functions
static.