Fix abort not being send when the state is idle
The spec clearly suggested that abort can be send while on idle state:

"AVDTP_ABORT_CMD can be sent or received in IDLE state. In the event that
an AVDTP_ABORT_CMD is received in IDLE state, ACP or INT shall reply with
an AVDTP_ABORT_RSP, no state change is required."

In fact abort is the only way to cancel set_configuration and open
commands.