Keep found LE device list to use when creating a device
Found devices list should be used to identify device types when
CreateDevice or CreatePairedDevice are called. During the device
creation the selected device will be classified as LE device if it
belongs to the device found list and if the Flags AD type matches the
GAP requirements. BR/EDR devices don't need to be tracked since the
default remote device type is Basic Rate.